Create Product folder under component and move Product.js there
